Added Cell Phone Records and SubDictionary Project
This project is composed of two parts, the Cell Phone Records and the Sub Dictionary programs. The Sub Dictionary program was worked on by my partner, while the Cell Phone Records program was worked on myself. 

The Cell Phone records program can be found under the SRC folder, and then in the Cell Phone records file, which will contain the java programs necessary to execute the project.